Github issue assigner webservice
This simple webservices listens to github issues webhooks and when they
are created, it automatically assignees one of the users on a list of
users as the assignee for that issue.

Uses a github personal token and github user list as environment
variable. It is configured to be readily deployed to heroku.